+++ ush/generate_FV3LAM_wflow.py
@@ -3,7 +3,6 @@
 import os
 import sys
 import subprocess
-import unittest
 import logging
 from multiprocessing import Process
 from textwrap import dedent
