+++ ush/set_gridparams_ESGgrid.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import os
-import unittest
 from datetime import datetime, timedelta
 
 from python_utils import (
@@ -74,43 +73,3 @@ def set_gridparams_ESGgrid(
         "NEG_NX_OF_DOM_WITH_WIDE_HALO": int(-(nx + 2 * halo_width)),
         "NEG_NY_OF_DOM_WITH_WIDE_HALO": int(-(ny + 2 * halo_width)),
     }
-
-
-class Testing(unittest.TestCase):
-    def test_set_gridparams_ESGgrid(self):
-
-        grid_parms = set_gridparams_ESGgrid(
-            lon_ctr=-97.5,
-            lat_ctr=38.5,
-            nx=1748,
-            ny=1038,
-            pazi=0.0,
-            halo_width=6,
-            delx=3000.0,
-            dely=3000.0,
-            constants=dict(
-                RADIUS_EARTH=6371200.0,
-                DEGS_PER_RADIAN=57.29577951308232087679,
-            ),
-        )
-
-        self.assertEqual(
-            list(grid_parms.values()),
-            [
-                -97.5,
-                38.5,
-                1748,
-                1038,
-                0.0,
-                6,
-                0.999,
-                0.013489400626196555,
-                0.013489400626196555,
-                -1760,
-                -1050,
-            ],
-        )
-
-    def setUp(self):
-        set_env_var("DEBUG", False)
-        set_env_var("VERBOSE", False)
